+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Fix crafting station requirements for Monstrum / Warfare / SouthsilArmor items.
+
 ## [0.5.0] - 2024-05-06
 
 ### Added